% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{H3N2_small}
\alias{H3N2_small}
\title{An integrated seurat object from an H3N2 infected experiment.}
\format{
An Seurat object
}
\source{
\url{https://www.ncbi.nlm.nih.gov/sra/?term=SRP239555}
}
\usage{
H3N2_small
}
\description{
An R-data containing an integrated seurat object with 
46 H3N2 infected cells and 46 H3N2 bystander cells.
}
\keyword{datasets}
