% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HALLMARK}
\alias{HALLMARK}
\title{An list object of MSIGDB HALLMARK dataset}
\format{
An list object
}
\source{
\url{http://www.gsea-msigdb.org/gsea/msigdb/collections.jsp}
}
\usage{
HALLMARK
}
\description{
An list object of MSIGDB HALLMARK dataset
}
\keyword{datasets}
