% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDensityPlot.R
\name{scDensityPlot}
\alias{scDensityPlot}
\title{scDensityPlot function}
\usage{
scDensityPlot(
  object = NULL,
  reduction = NULL,
  title = NULL,
  split.by = NULL,
  ncol = NULL,
  colors = NULL,
  digits = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Dimensional reduction to use, default reduction='umap'}

\item{title}{Figure title}

\item{split.by}{Name of a meta.data column to split plot by}

\item{ncol}{Number of columns for display the plots}

\item{colors}{Vector of colours to use for n-colour gradient}

\item{digits}{Integer representing the number of decimal places reserved for
the legend, default digits=3}
}
\value{
A ggplot object
}
\description{
Visualize density plot in low dimensional embedding
}
\examples{
scDensityPlot(object = H3N2_small,
  reduction = "umap",
  split.by = "sample",
  ncol = 2
)

}
