% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDGE.R
\name{scDGE}
\alias{scDGE}
\title{Differential gene expression analysis for each cluster
Perform differential analysis (DE) between two groups for each clusters}
\usage{
scDGE(
  object = object,
  comparison = c("condA", "condB"),
  group.by = NULL,
  min.cells = 20,
  min.pct = 0.1,
  logFC = 0.25,
  clusters = NULL
)
}
\arguments{
\item{object}{Seurat object.}

\item{comparison}{Vector of nominal variables from group.by. 
Eg., comparison=c('Normal', 'Tumor')}

\item{group.by}{Regroup cells before performing DGE, 
default group.by='group'}

\item{min.cells}{Minimum number of cells for DGE in any of the two groups, 
default min.cells=20}

\item{min.pct}{Genes expressed in any of the two groups with a minium cell
fraction (min.pct) are reserved for DGE, default min.pct=0.1}

\item{logFC}{A positive value to set the cutoff of logFC, default logFC=0.25}

\item{clusters}{Vector of clusters to perform DGE, default clusters=NULL,
which means all clusters will be evaluated}
}
\value{
DGE Table
}
\description{
Differential gene expression analysis for each cluster
Perform differential analysis (DE) between two groups for each clusters
}
\examples{
data("H3N2_small")
x <- scDGE(object = H3N2_small, 
  comparison = c("Infected", "Bystander"),
  group.by = "group", 
  min.cells = 10, 
  logFC = 0.25, 
  clusters = NULL
)
head(x)

}
\seealso{
\code{\link[Seurat]{FindMarkers}}
}
