% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGO.R
\name{scGO}
\alias{scGO}
\title{scGO}
\usage{
scGO(
  object = NULL,
  key = NULL,
  clusters = NULL,
  logFC = 0.25,
  only.pos = TRUE,
  reference = "human",
  extra = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{key}{Keyword where DGEs are stored}

\item{clusters}{A vector of clusters, default clusters=NULL}

\item{logFC}{A positive value to set the cutoff of logFC}

\item{only.pos}{Only up-regulated genes are taken into account, 
default only.pos=TRUE}

\item{reference}{Reference organism to set ['human' or 'mouse'], 
default reference = 'human'}

\item{extra}{Extra slot used to fetch DGEs}
}
\value{
A list object that stored GO results.
}
\description{
GO annotation for single-cell
}
\examples{
data("H3N2_small")
require(topGO)
x <- scGO(object = H3N2_small,
  key = "Infected_vs_Bystander", 
  logFC = 0.25,
  only.pos = FALSE,
  reference = "human"
)
head(x)

}
\references{
Inspired by Yuan et al, Methods in Molecular Biology (2019)
}
