% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGOBarPlot.R
\name{scGOBarPlot}
\alias{scGOBarPlot}
\title{scGOBarPlot visualization}
\usage{
scGOBarPlot(
  object = NULL,
  key = NULL,
  cluster = NULL,
  direction = "up",
  ont = "BP",
  top_n = 20,
  extra = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{key}{The 'scGO' results that stored in object@misc$key}

\item{cluster}{Cluster to presentation}

\item{direction}{Regulation direction to use, only 'up' or 'down'}

\item{ont}{Ontology to use, only 'BP', 'CC', or 'MF'}

\item{top_n}{Only top_n entries to plot}

\item{extra}{Extra slot used to fetch GOs}
}
\value{
A ggplot object
}
\description{
GO BarPlot visualization for each cluster
}
\examples{
scGOBarPlot(object = H3N2_small,
  key = "Infected_vs_Bystander.GO",
  ont = "BP",
  top_n = 6,
  direction = "up",
  cluster = "0"
)

}
