% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGODotPlot.R
\name{scGODotPlot}
\alias{scGODotPlot}
\title{scGODotPlot visualization}
\usage{
scGODotPlot(
  object = NULL,
  key = NULL,
  clusters = NULL,
  direction = "up",
  ont = "BP",
  padj.cutoff = 0.01,
  order.genes = FALSE,
  colors = NULL,
  font.size = 8,
  top_n = 5,
  extra = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{key}{The keyword that is stored in object@misc$key}

\item{clusters}{Vectors of clusters to display, default clusters=NULL 
means all clusters}

\item{direction}{Regulation direction to use, only 'up' or 'down' is allowed}

\item{ont}{Ontology to use, only 'BP', 'CC', or 'MF' is allowed}

\item{padj.cutoff}{The cutoff of adjusted Pvalue, default padj.cutoff=0.01}

\item{order.genes}{Order GO results by the number of significant genes, 
default is FALSE}

\item{colors}{Colors to specify, default cols = c('white', 'red')}

\item{font.size}{Font size of axis}

\item{top_n}{Only top_n entries to plot}

\item{extra}{Extra slot used to fetch GOs}
}
\value{
A ggplot object
}
\description{
scGODotPlot visualization for all clusters
}
\examples{
scGODotPlot(object = H3N2_small,
  key = "Infected_vs_Bystander.GO",
  ont = "BP",
  top_n = 6,
  direction = "up",
  cluster = "0",
  font.size = 8
)

}
