% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scIntegrate.R
\name{scIntegrate}
\alias{scIntegrate}
\title{scIntegrate}
\usage{
scIntegrate(
  object.list = NULL,
  object.names = NULL,
  nVariable = 2000,
  nPC = 30,
  resolution = 0.5,
  batch.rm = "harmony"
)
}
\arguments{
\item{object.list}{A list of seurat objects}

\item{object.names}{An array of seurat object names}

\item{nVariable}{Number of features to select as top variable features}

\item{nPC}{Number of PCs to use}

\item{resolution}{Resolution parameter to set. Default resolution=0.5}

\item{batch.rm}{Remove batch effect with 'seurat' or 'harmony'. 
Default batch.rm='harmony'}
}
\value{
Seurat object.
}
\description{
Integrate seurat objects into one
}
\examples{
Bystander <- scRead(sample_name = "Bystander", 
                    data_dir = system.file(
                      "extdata/H3N2_10X_matrix/Bystander/", 
                    package="Yeskit"), 
                    gene_column = 2, project_name = "H3N2", 
                    group_name = "Bystander", 
                    meta_file = system.file(
                      "extdata/H3N2_10X_matrix/Bystander/microbes.tsv", 
                      package="Yeskit")
                    )
Infected <- scRead(sample_name = "Infected", 
                    data_dir = system.file(
                      "extdata/H3N2_10X_matrix/Infected/", 
                    package="Yeskit"), 
                    gene_column = 2, project_name = "H3N2", 
                    group_name = "Infected", 
                    meta_file = system.file(
                      "extdata/H3N2_10X_matrix/Infected/microbes.tsv", 
                      package="Yeskit")
                    )
Integrated <- scIntegrate(object.list = list(Bystander, Infected), 
  object.names = c("Bystander", "Infected"),
  batch.rm = "harmony",
  resolution = 0.7
)

}
\references{
Inspired by Stuart and Butler et al, Cell (2019)
}
