% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMSigdbScoring.R
\name{scMsigdbScoring}
\alias{scMsigdbScoring}
\title{scMsigdbScoring}
\usage{
scMsigdbScoring(
  object = NULL,
  species = "human",
  category = NULL,
  genesets = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{species}{Species name. Such as "human" or "mouse"}

\item{category}{String of MSigDB category to use, 
default category='H' stands for 'hallmark gene sets'}

\item{genesets}{Vector of genesets under the specific category of 
MSIGDB, such as "APOPTOSIS". Default genesets=NULL, which means 
all genesets under the specified category will be used}
}
\value{
Seurat object
}
\description{
MSigdb signature scoring for single-cells
}
\examples{
data("H3N2_small")
x <- scMsigdbScoring(object = H3N2_small, 
  species = "human", category = "H",
  genesets = c("HALLMARK_INFLAMMATORY_RESPONSE")
)
head(x)

}
