% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scOne.R
\name{scOne}
\alias{scOne}
\title{single-cell analysis of one sample}
\usage{
scOne(
  object = NULL,
  nFeatures = 2000,
  nPC = 30,
  resolution = 0.5,
  perplexity = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{nFeatures}{Number of variable features to use}

\item{nPC}{Number of principal components to use}

\item{resolution}{Value of the resolution parameter}

\item{perplexity}{Value of the perplexity parameter, set between 5 and 50}
}
\value{
Seurat object
}
\description{
A wrapper of the standard workflow of Seurat
}
\examples{
Bystander <- scRead(sample_name = "Bystander", 
                    data_dir = system.file(
                      "extdata/H3N2_10X_matrix/Bystander/", 
                    package="Yeskit"), 
                    gene_column = 2, project_name = "H3N2", 
                    group_name = "Bystander", 
                    meta_file = system.file(
                      "extdata/H3N2_10X_matrix/Bystander/microbes.tsv", 
                      package="Yeskit")
                    )
Infected <- scRead(sample_name = "Infected", 
                    data_dir = system.file(
                      "extdata/H3N2_10X_matrix/Infected/", 
                    package="Yeskit"), 
                    gene_column = 2, project_name = "H3N2", 
                    group_name = "Infected", 
                    meta_file = system.file(
                      "extdata/H3N2_10X_matrix/Infected/microbes.tsv", 
                      package="Yeskit")
                    )
Integrated <- merge(Bystander, Infected)
Integrated <- scOne(object = Integrated, 
  nFeatures = 2000,
  nPC = 30,
  resolution = 0.7
)

}
\references{
Inspired by Stuart and Butler et al, Cell (2019)
}
