% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPathogenDGE.R
\name{scPathogenDGE}
\alias{scPathogenDGE}
\title{Find differentially expressed genes by species for each clusters}
\usage{
scPathogenDGE(
  object = NULL,
  species.by = NULL,
  clusters = NULL,
  min.cells = 20,
  min.pct = 0.1,
  logFC = 0.25
)
}
\arguments{
\item{object}{Seurat object}

\item{species.by}{String of species to divid cells into 
two groups (Positive / Negative)}

\item{clusters}{Clusters specified for DGE, default clusters=NULL, means
all clusters will be evaluated}

\item{min.cells}{Minimum number of cells for DGE in any of the two groups,
default min.cells=20}

\item{min.pct}{Genes expressed in any of the two groups with a minium cell 
fraction (min.pct) are reserved for DGE, default min.pct=0.1}

\item{logFC}{A positive value to set the cutoff of logFC, default logFC=0.25}
}
\value{
DGE table
}
\description{
Find differentially expressed genes by species for each clusters
}
\examples{
data("H3N2_small")
x <- scPathogenDGE(object = H3N2_small, 
  species.by = "H3N2",
  min.cells = 5,
  min.pct = 0.1,
  logFC = 0.25,
)
head(x)

}
\references{
Inspired by Stuart and Butler et al, Cell (2019)
}
