% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPathogenGO.R
\name{scPathogenGO}
\alias{scPathogenGO}
\title{scPathogenGO annotation}
\usage{
scPathogenGO(
  object = NULL,
  key = NULL,
  species = NULL,
  clusters = NULL,
  logFC = 0.25,
  only.pos = TRUE,
  reference = "human"
)
}
\arguments{
\item{object}{Seurat object}

\item{key}{The DGEs are stored in object@misc$key}

\item{species}{Species which is used to divid cells into two groups}

\item{clusters}{A vector of clusters, default all clusters will be used}

\item{logFC}{A positive value to set the cutoff of logFC}

\item{only.pos}{Only up-regulated genes are taken into account, 
default only.pos=TRUE}

\item{reference}{Reference organism to set ['human' or 'mouse'], 
default reference = 'human'}
}
\value{
Table that stores scPathogenGO results
}
\description{
GO annotation for scPathogenDGE results
}
\examples{
data("H3N2_small")
require(topGO)
x <- scPathogenGO(object = H3N2_small,
  key = "H3N2", 
  species = "H3N2",
  logFC = 0.25,
  only.pos = FALSE,
  reference = "human"
)
head(x)

}
\references{
Inspired by Yuan et al, Methods in Molecular Biology (2019)
}
