% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scScoreDimPlot.R
\name{scScoreDimPlot}
\alias{scScoreDimPlot}
\title{scScoreDimPlot}
\usage{
scScoreDimPlot(
  object = NULL,
  signature = NULL,
  reduction = NULL,
  cols = NULL,
  pt.size = NULL,
  split.by = NULL,
  title = NULL,
  ncol = NULL,
  raster = TRUE,
  scale = TRUE,
  col.min = NA,
  col.max = NA
)
}
\arguments{
\item{object}{Seurat object}

\item{signature}{Name of one gene set}

\item{reduction}{Which dimensionality reduction to use}

\item{cols}{Colors to plot}

\item{pt.size}{Adjust point size to plot, default pt.size=0.5}

\item{split.by}{Name of a metadata column to split plot by}

\item{title}{Title of the plot}

\item{ncol}{Number of columns for display the plots}

\item{raster}{Convert points to raster format, default is TRUE}

\item{scale}{Determine whether to scale the data, default is TRUE}

\item{col.min}{Minimum scaled average score threshold 
(smaller values will be set to this)}

\item{col.max}{Maximum scaled average score threshold 
(larger values will be set to this)}
}
\value{
A ggplot object
}
\description{
MSigdb scoring DimPlot for single-cell clusters
}
\examples{
data("H3N2_small")
scScoreDimPlot(object = H3N2_small,
  signature = "HALLMARK_INFLAMMATORY_RESPONSE",
  reduction = "umap",
  cols = NULL,
  split.by = "sample",
  ncol = 2,
  pt.size = 1 
)

}
