% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVizMeta.R
\name{scVizMeta}
\alias{scVizMeta}
\title{scVizMeta function}
\usage{
scVizMeta(
  object = NULL,
  signature = NULL,
  pt.size = NULL,
  reduction = NULL,
  title = NULL,
  split.by = NULL,
  ncol = NULL,
  raster = TRUE,
  palette = "YlOrRd",
  interval = c(Max = 1e+05, Hyper = 10000, Abundant = 1000, Large = 500, 
    Medium = 100, Small = 10, Single = 1, None = 0)
)
}
\arguments{
\item{object}{Seurat object}

\item{signature}{Name of one gene set}

\item{pt.size}{Adjust point size to plot, default pt.size=0.5}

\item{reduction}{Which dimensionality reduction to use, default is 'umap'}

\item{title}{Text for the plot title, default title=NULL}

\item{split.by}{Name of a metadata column to split plot by}

\item{ncol}{Number of columns for display the plots}

\item{raster}{Convert points to raster format, default is TRUE}

\item{palette}{A palette name from RColorBrewer package}

\item{interval}{set MetaData intervals to plot}
}
\value{
A ggplot object
}
\description{
Visualize MetaData in low dimensional embedding
}
\details{
visualize data from MetaData slot
}
\examples{
data("H3N2_small")
scVizMeta(object = H3N2_small,
  reduction = "umap",
  signature = "H3N2",
  split.by = "sample",
  pt.size = 1,
  interval = c(
    Abundant = 1000, 
    Large = 500, 
    Medium = 100, 
    Small = 10, 
    Single = 1, 
    None = 0
  )
)

}
\references{
Inspired by Borcherding et al, F1000Research (2020)
}
