% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scVolcanoPlot.R
\name{scVolcanoPlot}
\alias{scVolcanoPlot}
\title{scVolcanoPlot}
\usage{
scVolcanoPlot(
  object = NULL,
  key = NULL,
  cluster = NULL,
  padj = 0.05,
  logFC = 0.585,
  cols = NULL,
  pt.size = NULL,
  species.by = NULL,
  top_n = NULL,
  labels = NULL,
  label.size = NULL,
  box = TRUE,
  raster = TRUE
)
}
\arguments{
\item{object}{A seurat object.}

\item{key}{The DE matrix are stored in object@misc$key}

\item{cluster}{Cluster specified by the user}

\item{padj}{Cutoff of adjusted pvalue}

\item{logFC}{Log2FoldChange in DGEs}

\item{cols}{Colours to pain the dots, 
default cols=c('#808080', '#810F7C', '#006D2C', '#000000')}

\item{pt.size}{Size of the dots, default pt.size=1}

\item{species.by}{Species which is used to divid cells into two groups}

\item{top_n}{Number of up- and down-regulated genes to show}

\item{labels}{An array of gene names to show}

\item{label.size}{Font size of the labels, default label.size=2.5}

\item{box}{Draws a rectangle behind the labels, default is TRUE}

\item{raster}{Convert points to raster format, default is TRUE}
}
\value{
A ggplot object
}
\description{
Volcano Plot for DGEs
}
\examples{
data("H3N2_small")
scVolcanoPlot(object = H3N2_small,
  key = "Infected_vs_Bystander",
  cluster = "0",
  top_n = 5
)

}
